move all cnn_lstm components to examples
