+++ examples/cnn_lstm/cnn_encoder.py
