+++ examples/cnn_lstm/lstm_encoder.py
