+++ examples/cnn_lstm/test/test_cnn_encoder.py
@@ -7,9 +7,9 @@
 import unittest
 
 import torch
+from examples.cnn_lstm.cnn_encoder import CNNEncoder
 from test.test_utils import assert_expected, set_rng_seed
 from torch import nn, Tensor
-from torchmultimodal.modules.encoders.cnn_encoder import CNNEncoder
 
 
 class TestCNNEncoder(unittest.TestCase):
